Remove revoked SSH key pairs
I'm an idiot. I committed SSH key pairs and pushed to a public
repository without thinking. There was a pair for github, gitlab,
StarLeaf and Newcastle University. I've revoked the first two and
contacted the last two to inform them of the leak. (I hope SL and NCL no
longer allow my user to do anything, but that's not the point.)